// src/api/generated/model/userResponseRoles.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */

export type UserResponseRoles = typeof UserResponseRoles[keyof typeof UserResponseRoles];


export const UserResponseRoles = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  DRIVER: 'DRIVER',
  BACKOFFICE: 'BACKOFFICE',
  PRE_AUTH: 'PRE_AUTH',
} as const;